+++ libavformat/spcm.c
@@ -35,13 +35,13 @@ static int read_probe(const AVProbeData *p)
     if (AV_RL32(p->buf+4) != 0x00010000)
         return 0;
 
-    if (AV_RL32(p->buf+8) <= 0)
+    if ((int32_t)AV_RL32(p->buf+8) <= 0)
         return 0;
 
-    if (AV_RL32(p->buf+12) <= 0)
+    if ((int32_t)AV_RL32(p->buf+12) <= 0)
         return 0;
 
-    if (AV_RL32(p->buf+16) <= 0)
+    if ((int32_t)AV_RL32(p->buf+16) <= 0)
         return 0;
 
     return AVPROBE_SCORE_MAX;
